fix: la-136 split requestVerificationCode business logic
